Update `DocumentCollection` sync check
Some linked documents do not have an edition that is in a visible state
e.g. only superseded. This is predominantly when they have been
'deleted' but there are other examples.

This commit checks for the presence of a 'visible' edition on the
docuement rather than just that the document has been deleted. This
should be more robust and catch all cases rather than just the
'legitimate' deletion case. If the linked document does not have a
visible edition it is excluded from the check.