Somewhat crude Link Checker API implementation into monthly check
This integrates Link Checker API with the Whitehall monthly check.
Unfortunately the only way we're going to know how well it works is to
run it with a full set of data - which last time took over 24 hours to
run.

The things to be concerned about with this implementation is:

- A loop runs until all link checks are considered completed, if
  any of these never reach a completed status then the loop will
  continue forever.
- Any single non successful calls to Link Checker API will abort the
  entire process, which seems particularly painful in a multi hour
  running process.
- There isn't any easing off of requests, so if one edition is taking a
  long time, we might end up hitting Link Checker API multiple times per
  second
- As this will flood Link Checker API with requests it may block ability
  to do other link checking while this has filled the queue.
- We load every public edition (~ 180,000) into an array of objects,
  this might be significantly painful memory wise. We probably don't
  need the full objects.

Note: This also adds a fallback for an edition that responds to
worldwide_organisations but doesn't have one and thus returns nil,
instead they return first organisation.